refactoring of events, added an endpoint to get a list of regions of Ukraine
