+++ project/cities/constants/errors.py
@@ -1,0 +1,3 @@
+NOTHING_FOUND_FOR_USER_REQUEST_ERROR: dict[str, str] = {
+    "error": "nothing_found_for_request"
+}
