+++ project/events/serializers.py
@@ -4,13 +4,11 @@ from typing import Any, Union
 from authentication.models import User
 from authentication.serializers import (
     EventUsersSerializer,
+)
+from cities.serializers import (
     PlaceSerializer,
 )
 from config.exceptions import _404
-from django.core.validators import (
-    MaxValueValidator,
-    MinValueValidator,
-)
 from events.constants.response_error import (
     ALREADY_IN_EVENT_FANS_LIST_ERROR,
     ALREADY_IN_EVENT_MEMBERS_LIST_ERROR,
@@ -132,7 +130,12 @@ class InvitesToEventListSerializer(serializers.ModelSerializer):
 
     class Meta:
         model: InviteToEvent = InviteToEvent
-        fields: Union[str, list[str]] = ["id", "time_created", "event", "sender"]
+        fields: Union[str, list[str]] = [
+            "id",
+            "time_created",
+            "event",
+            "sender",
+        ]
 
 
 class DeleteEventsSerializer(serializers.Serializer):
@@ -250,34 +253,4 @@ class BulkAcceptOrDeclineRequestToParticipationSerializer(serializers.Serializer
         fields: Union[str, list[str]] = [
             "ids",
             "type",
-        ]
-
-
-class GetCoordinatesByPlaceNameSerializer(serializers.Serializer):
-    place_name: bool = serializers.CharField(max_length=255)
-
-    class Meta:
-        fields: Union[str, list[str]] = [
-            "place_name",
-        ]
-
-
-class GetPlaceNameByCoordinatesSerializer(serializers.Serializer):
-    lat: float = serializers.FloatField(
-        validators=[
-            MinValueValidator(-90),
-            MaxValueValidator(90),
-        ]
-    )
-    lon: float = serializers.FloatField(
-        validators=[
-            MinValueValidator(-180),
-            MaxValueValidator(180),
-        ]
-    )
-
-    class Meta:
-        fields: Union[str, list[str]] = [
-            "lat",
-            "lon",
         ]
