+++ project/events/urls.py
@@ -14,9 +14,7 @@ from events.views import (
     EventsRelevantList,
     FanJoinToEvent,
     FanLeaveFromEvent,
-    GetCoordinatesByPlaceName,
     GetEvent,
-    GetPlaceNameByCoordinates,
     InvitesToEventList,
     InviteUserToEvent,
     JoinToEvent,
@@ -104,15 +102,5 @@ urlpatterns: list[Union[URLResolver, URLPattern]] = [
         "client/accept/or/decline/invites/to/events",
         BulkAcceptOrDeclineInvitesToEvent.as_view(),
         name="accept-decline-invites-to-event",
-    ),
-    path(
-        "client/get/coordinates/by/place/name",
-        GetCoordinatesByPlaceName.as_view(),
-        name="get-coordinates-by-place-name",
-    ),
-    path(
-        "client/get/place/name/by/coordinates",
-        GetPlaceNameByCoordinates.as_view(),
-        name="get-place-name-by-coordinates",
     ),
 ]
